pyproject.toml
[project]
name = "system-card-eval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "adjusttext>=1.3.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "openai>=2.30.0",
    "pandas>=3.0.2",
    "pymupdf>=1.27.2.2",
    "pymupdf4llm>=1.27.2.2",
    "python-dotenv>=1.2.2",
]
